Shapes: enable shapes in GeoJSON layer as well
